pyproject.toml
[tool.poetry]
name = "advent2021"
version = "0.1.0"
description = ""
authors = ["Kevin Wilson <khwilson@gmail.com>"]
packages = [
    { include = "advent", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
click = "^8.0.3"
numpy = "^1.21.4"
scipy = "^1.7.3"
matplotlib = "^3.5.0"
seaborn = "^0.11.2"
pandas = "^1.3.4"
python-dotenv = "^0.19.2"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.11b1"
pylint = "^2.6.0"
isort = "^5.10.1"
nbqa = "^1.2.2"
mypy = "^0.910"
ipdb = "^0.13.4"

[tool.poetry.scripts]
advent = "advent.cli:run_solution"
advent-download = "advent.cli:download_data"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
